pyproject.toml
[project]
name = "rss-reader"
version = "0.1.0"
description = "RSS reader with LLM summaries and daily email digest"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.1",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.4",
  "feedparser>=6.0.11",
  "apscheduler>=3.10.4",
  "jinja2>=3.1.4",
  "httpx>=0.27.0",
  "PyYAML>=6.0.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
